requirements.txt
python-telegram-bot==20.7
requests
pymongo
python-dotenv
python-dateutil